Update AD2SmartThings_v3_0.ino
I corrected this earlier because it threw an error. Now that is the syntax is correct - it sends [CODE] along with the command to send to the AD2Pi. As far as I can tell, the AD2Pi doesn't like that and I'm assuming that it's expecting just the numeric cmd.